cls/data_io/suitcase_nxstxm/requirements-dev.txt
# These are required for developing the package (running the tests, building
# the documentation) but not necessarily required for _using_ it.
codecov
coverage
flake8
# TEMPORARY UNTIL https://github.com/NSLS-II/ophyd/pull/682 IS RELEASED
git+git://github.com/awalter-bnl/ophyd@add-configure
pytest >=3.9
sphinx
suitcase-utils[test_fixtures] >=0.1.0rc2
# These are dependencies of various sphinx extensions for documentation.
ipython
matplotlib
numpydoc
sphinx-copybutton
sphinx_rtd_theme
